Moved PottsCell make, state, and module tests into PottsCellStemTest. Made mock implementation of PottsCell for PottsCellTest
